requirements.txt
certifi==2020.12.5
chardet==3.0.4
idna==2.5
requests2==2.16.0
urllib3==1.21.1
